Use size_t instead of u_int for total memory.
In testing, when using buf_num over 420,000 or so, with the default
clusternumbers, we run into integer overflow and segfault.

In our environment, we have 40 core servers (20+20) with the nics using
40 tx/rx rings, with the ring slot size set to 4096. We have to set the
buf_num to about 750000 in order to allocate enough memory for our
application to load with around 100000 extra buffers.

This segfaults without this change.

This is basically adapted from issue #602. I left the max buffers at the
default of 1_000_000 and changed some %lu and %ld to %zu %zd

+++ utils/ctrl-api-test.c
@@ -206,7 +206,7 @@ port_info_get(struct TestContext *ctx)
 		perror("ioctl(/dev/netmap, NIOCCTRL, PORT_INFO_GET)");
 		return ret;
 	}
-	printf("nr_memsize %llu\n", (unsigned long long)req.nr_memsize);
+	printf("nr_memsize %zu\n", req.nr_memsize);
 	printf("nr_tx_slots %u\n", req.nr_tx_slots);
 	printf("nr_rx_slots %u\n", req.nr_rx_slots);
 	printf("nr_tx_rings %u\n", req.nr_tx_rings);
@@ -265,7 +265,7 @@ port_register(struct TestContext *ctx)
 		return ret;
 	}
 	printf("nr_offset 0x%llx\n", (unsigned long long)req.nr_offset);
-	printf("nr_memsize %llu\n", (unsigned long long)req.nr_memsize);
+	printf("nr_memsize %zu\n", req.nr_memsize);
 	printf("nr_tx_slots %u\n", req.nr_tx_slots);
 	printf("nr_rx_slots %u\n", req.nr_rx_slots);
 	printf("nr_tx_rings %u\n", req.nr_tx_rings);
@@ -339,7 +339,7 @@ niocregif(struct TestContext *ctx, int netmap_api)
 	}
 
 	printf("nr_offset 0x%x\n", req.nr_offset);
-	printf("nr_memsize  %u\n", req.nr_memsize);
+	printf("nr_memsize  %zu\n", req.nr_memsize);
 	printf("nr_tx_slots %u\n", req.nr_tx_slots);
 	printf("nr_rx_slots %u\n", req.nr_rx_slots);
 	printf("nr_tx_rings %u\n", req.nr_tx_rings);
